requirements.txt
torch>=1.4.0
nltk
pandas
PyYAML
matplotlib
tqdm
wget
youtokentome
